src/Inserting_value_in_bigQuery/requirements.txt
#requirements.txt
pandas
geopandas
matplotlib
pysal
geojson
requests
python-dotenv
google-cloud-bigquery
google-cloud-storage
google-cloud-bigquery[pandas] 
pymongo
